others/graphql/graphene_gjango/requirements.txt
aniso8601==9.0.1
arrow==1.2.3
asgiref==3.6.0
Django==4.1.4
django-classy-tags==4.0.0
django-cms==3.11.1
django-formtools==2.4
django-phonenumber-field==7.0.1
django-phonenumbers==1.0.1
django-seed==0.3.1
django-sekizai==4.0.0
django-treebeard==4.5.1
djangocms-admin-style==3.2.1
djangocodemirror==2.1.0
djangorestframework==3.14.0
emencia-cms-snippet==2.4.2
Faker==15.3.4
graphene==3.2.1
graphene-django==3.0.0
graphene-django-cud==0.10.0
graphene-file-upload==1.3.0
graphql-core==3.2.3
graphql-relay==3.2.0
install==1.3.5
Jinja2==3.1.2
jinja2-time==0.2.0
make==0.1.6.post2
MarkupSafe==2.1.1
phonenumbers==8.13.3
promise==2.3
psycopg2-binary==2.9.5
python-dateutil==2.8.2
pytz==2022.7
six==1.16.0
sqlparse==0.4.3
text-unidecode==1.3
toposort==1.7
